# core/docs/swagger.yaml
definitions:
...
  apprun_modules_auth_service.RegisterRequest:
    properties:
      email:
        description: 用户邮箱（必填）
        example: user@example.com
        type: string
      gender:
        description: 性别（可选：0=未知,1=男,2=女）
        example: 1
        type: integer
      language:
        description: 语言（可选）
        example: zh-CN
        type: string
      nickname:
        description: 昵称（可选）
        example: John
        type: string
      password:
        description: 密码（必填，8+字符，含大小写和数字）
        example: SecurePass123
        type: string
      phone:
        description: 手机号（可选）
        example: +86-13800138000
        type: string
      timezone:
        description: 时区（可选）
        example: Asia/Shanghai
        type: string
      username:
        description: 用户名（可选，3-64字符）
        example: john_doe
        type: string
    required:
    - email
    - password
    type: object
  apprun_modules_auth_service.RegisterResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        description: Public UUID (exposed as "id" for external API)
        type: string
      nickname:
        type: string
      status:
        type: integer
      username:
        type: string
    type: object
...
  apprun_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  apprun_pkg_response.PaginationInfo:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  apprun_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        $ref: '#/definitions/apprun_pkg_response.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
  ent.Project:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.ProjectEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProjectQuery when eager-loading is set.
      id:
        description: |-
          ID of the ent.
          Internal database primary key for foreign key relationships
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      owner_id:
        description: Project owner user_id, foreign key to users.id
        type: integer
      status:
        description: 'Status: 0-disabled, 1-enabled, 2-archived'
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      uuid:
        description: External UUID for API usage in REST paths
        type: string
    type: object
  ent.ProjectEdges:
    properties:
      members:
        description: Members holds the value of the members edge.
        items:
          $ref: '#/definitions/ent.ProjectMember'
        type: array
      owner:
        allOf:
        - $ref: '#/definitions/ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  ent.ProjectMember:
    properties:
      edges:
        allOf:
        - $ref: '#/definitions/ent.ProjectMemberEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProjectMemberQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      joined_at:
        description: JoinedAt holds the value of the "joined_at" field.
        type: string
      project_id:
        description: Foreign key to projects.id
        type: integer
      role:
        description: 'Role: owner, admin, member, viewer'
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      user_id:
        description: Foreign key to users.id
        type: integer
    type: object
  ent.ProjectMemberEdges:
    properties:
      project:
        allOf:
        - $ref: '#/definitions/ent.Project'
        description: Project holds the value of the project edge.
      user:
        allOf:
        - $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
    type: object
  ent.Servers:
    properties:
      edges:
        allOf:
        - $ref: '#/definitions/ent.ServersEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ServersQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      ip:
        description: IP holds the value of the "ip" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.ServersEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  ent.User:
    properties:
      avatar:
        description: Avatar URL
        type: string
      created_at:
        description: Account creation timestamp
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      email:
        description: Email address for login (required, unique)
        type: string
      gender:
        description: 'Gender: 0-Unknown, 1-Male, 2-Female'
        type: integer
      id:
        description: |-
          ID of the ent.
          Internal database primary key (not exposed in API)
        type: integer
      language:
        description: Preferred language (e.g., zh-CN, en-US)
        type: string
      last_login_at:
        description: Last login timestamp
        type: string
      last_login_ip:
        description: Last login IP address (supports IPv6)
        type: string
      nickname:
        description: Display name / nickname
        type: string
      phone:
        description: Phone number
        type: string
      signature:
        description: User signature / bio
        type: string
      status:
        description: 'Account status: 0-Disabled, 1-Active'
        type: integer
      timezone:
        description: User timezone (e.g., Asia/Shanghai)
        type: string
      updated_at:
        description: Last update timestamp
        type: string
      username:
        description: Username for login (optional, alphanumeric + underscore)
        type: string
      uuid:
        description: Public UUID for external API reference (immutable, globally unique)
        type: string
    type: object
  ent.UserEdges:
    properties:
      owned_projects:
        description: OwnedProjects holds the value of the owned_projects edge.
        items:
          $ref: '#/definitions/ent.Project'
        type: array
      project_memberships:
        description: ProjectMemberships holds the value of the project_memberships
          edge.
        items:
          $ref: '#/definitions/ent.ProjectMember'
        type: array
      servers:
        description: Servers holds the value of the servers edge.
        items:
          $ref: '#/definitions/ent.Servers'
        type: array
    type: object
...
  modules_auth_handler.AddMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: integer
    required:
    - role
    - user_id
    type: object
...
  modules_auth_handler.MemberListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/ent.ProjectMember'
        type: array
      pagination:
        $ref: '#/definitions/apprun_pkg_response.PaginationInfo'
    type: object
...
  modules_auth_handler.UpdateRoleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
...
info:
  contact:
    email: support@websoft9.com
    name: API Support
  description: AppRun Platform REST API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AppRun API
  version: "1.0"
paths:
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apprun_modules_auth_service.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.RegisterResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "409":
          description: Email or username already exists
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: Register new user
      tags:
      - auth
...
  /api/projects/{project_id}/members:
    get:
      description: Get a list of all members in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/modules_auth_handler.MemberListResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - rbac
    post:
      consumes:
      - application/json
      description: Add a user to a project with a specific role (requires admin permission)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Member info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modules_auth_handler.AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ent.ProjectMember'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Add a member to a project
      tags:
      - rbac
  /api/projects/{project_id}/members/{member_id}:
    delete:
      description: Remove a user from a project (requires admin permission)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove project member
      tags:
      - rbac
    put:
      consumes:
      - application/json
      description: Update a project member's role (requires admin permission)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: integer
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modules_auth_handler.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ent.ProjectMember'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - rbac
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"